fix(sdk-coin-ada): remove explain transaction from lib
STLX-19723
